requirements.txt
anthropic>=0.49.0
agentscope>=1.0.18
boto3>=1.42.0
PyGithub>=2.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
python-telegram-bot==21.5
